Trait rusoto_ecr::Ecr

pub trait Ecr {
    fn batch_check_layer_availability(
        &self, 
        input: BatchCheckLayerAvailabilityRequest
    ) -> RusotoFuture<BatchCheckLayerAvailabilityResponse, BatchCheckLayerAvailabilityError>;

    fn batch_delete_image(
        &self, 
        input: BatchDeleteImageRequest
    ) -> RusotoFuture<BatchDeleteImageResponse, BatchDeleteImageError>;

    fn batch_get_image(
        &self, 
        input: BatchGetImageRequest
    ) -> RusotoFuture<BatchGetImageResponse, BatchGetImageError>;

    fn complete_layer_upload(
        &self, 
        input: CompleteLayerUploadRequest
    ) -> RusotoFuture<CompleteLayerUploadResponse, CompleteLayerUploadError>;

    fn create_repository(
        &self, 
        input: CreateRepositoryRequest
    ) -> RusotoFuture<CreateRepositoryResponse, CreateRepositoryError>;

    fn delete_lifecycle_policy(
        &self, 
        input: DeleteLifecyclePolicyRequest
    ) -> RusotoFuture<DeleteLifecyclePolicyResponse, DeleteLifecyclePolicyError>;

    fn delete_repository(
        &self, 
        input: DeleteRepositoryRequest
    ) -> RusotoFuture<DeleteRepositoryResponse, DeleteRepositoryError>;

    fn delete_repository_policy(
        &self, 
        input: DeleteRepositoryPolicyRequest
    ) -> RusotoFuture<DeleteRepositoryPolicyResponse, DeleteRepositoryPolicyError>;

    fn describe_images(
        &self, 
        input: DescribeImagesRequest
    ) -> RusotoFuture<DescribeImagesResponse, DescribeImagesError>;

    fn describe_repositories(
        &self, 
        input: DescribeRepositoriesRequest
    ) -> RusotoFuture<DescribeRepositoriesResponse, DescribeRepositoriesError>;

    fn get_authorization_token(
        &self, 
        input: GetAuthorizationTokenRequest
    ) -> RusotoFuture<GetAuthorizationTokenResponse, GetAuthorizationTokenError>;

    fn get_download_url_for_layer(
        &self, 
        input: GetDownloadUrlForLayerRequest
    ) -> RusotoFuture<GetDownloadUrlForLayerResponse, GetDownloadUrlForLayerError>;

    fn get_lifecycle_policy(
        &self, 
        input: GetLifecyclePolicyRequest
    ) -> RusotoFuture<GetLifecyclePolicyResponse, GetLifecyclePolicyError>;

    fn get_lifecycle_policy_preview(
        &self, 
        input: GetLifecyclePolicyPreviewRequest
    ) -> RusotoFuture<GetLifecyclePolicyPreviewResponse, GetLifecyclePolicyPreviewError>;

    fn get_repository_policy(
        &self, 
        input: GetRepositoryPolicyRequest
    ) -> RusotoFuture<GetRepositoryPolicyResponse, GetRepositoryPolicyError>;

    fn initiate_layer_upload(
        &self, 
        input: InitiateLayerUploadRequest
    ) -> RusotoFuture<InitiateLayerUploadResponse, InitiateLayerUploadError>;

    fn list_images(
        &self, 
        input: ListImagesRequest
    ) -> RusotoFuture<ListImagesResponse, ListImagesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn put_image(
        &self, 
        input: PutImageRequest
    ) -> RusotoFuture<PutImageResponse, PutImageError>;

    fn put_lifecycle_policy(
        &self, 
        input: PutLifecyclePolicyRequest
    ) -> RusotoFuture<PutLifecyclePolicyResponse, PutLifecyclePolicyError>;

    fn set_repository_policy(
        &self, 
        input: SetRepositoryPolicyRequest
    ) -> RusotoFuture<SetRepositoryPolicyResponse, SetRepositoryPolicyError>;

    fn start_lifecycle_policy_preview(
        &self, 
        input: StartLifecyclePolicyPreviewRequest
    ) -> RusotoFuture<StartLifecyclePolicyPreviewResponse, StartLifecyclePolicyPreviewError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn upload_layer_part(
        &self, 
        input: UploadLayerPartRequest
    ) -> RusotoFuture<UploadLayerPartResponse, UploadLayerPartError>;
}

Trait representing the capabilities of the Amazon ECR API. Amazon ECR clients implement this trait.

Required methods

fn batch_check_layer_availability(
&self,
input: BatchCheckLayerAvailabilityRequest
) -> RusotoFuture<BatchCheckLayerAvailabilityResponse, BatchCheckLayerAvailabilityError>

Check the availability of multiple image layers in a specified registry and repository.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

fn batch_delete_image(
&self,
input: BatchDeleteImageRequest
) -> RusotoFuture<BatchDeleteImageResponse, BatchDeleteImageError>

Deletes a list of specified images within a specified repository. Images are specified with either imageTag or imageDigest.

You can remove a tag from an image by specifying the image's tag in your request. When you remove the last tag from an image, the image is deleted from your repository.

You can completely delete an image (and all of its tags) by specifying the image's digest in your request.

fn batch_get_image(
&self,
input: BatchGetImageRequest
) -> RusotoFuture<BatchGetImageResponse, BatchGetImageError>

Gets detailed information for specified images within a specified repository. Images are specified with either imageTag or imageDigest.

fn complete_layer_upload(
&self,
input: CompleteLayerUploadRequest
) -> RusotoFuture<CompleteLayerUploadResponse, CompleteLayerUploadError>

Informs Amazon ECR that the image layer upload has completed for a specified registry, repository name, and upload ID. You can optionally provide a sha256 digest of the image layer for data validation purposes.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

fn create_repository(
&self,
input: CreateRepositoryRequest
) -> RusotoFuture<CreateRepositoryResponse, CreateRepositoryError>

Creates an image repository.

fn delete_lifecycle_policy(
&self,
input: DeleteLifecyclePolicyRequest
) -> RusotoFuture<DeleteLifecyclePolicyResponse, DeleteLifecyclePolicyError>

Deletes the specified lifecycle policy.

fn delete_repository(
&self,
input: DeleteRepositoryRequest
) -> RusotoFuture<DeleteRepositoryResponse, DeleteRepositoryError>

Deletes an existing image repository. If a repository contains images, you must use the force option to delete it.

fn delete_repository_policy(
&self,
input: DeleteRepositoryPolicyRequest
) -> RusotoFuture<DeleteRepositoryPolicyResponse, DeleteRepositoryPolicyError>

Deletes the repository policy from a specified repository.

fn describe_images(
&self,
input: DescribeImagesRequest
) -> RusotoFuture<DescribeImagesResponse, DescribeImagesError>

Returns metadata about the images in a repository, including image size, image tags, and creation date.

Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

fn describe_repositories(
&self,
input: DescribeRepositoriesRequest
) -> RusotoFuture<DescribeRepositoriesResponse, DescribeRepositoriesError>

Describes image repositories in a registry.

fn get_authorization_token(
&self,
input: GetAuthorizationTokenRequest
) -> RusotoFuture<GetAuthorizationTokenResponse, GetAuthorizationTokenError>

Retrieves a token that is valid for a specified registry for 12 hours. This command allows you to use the docker CLI to push and pull images with Amazon ECR. If you do not specify a registry, the default registry is assumed.

The authorizationToken returned for each registry specified is a base64 encoded string that can be decoded and used in a docker login command to authenticate to a registry. The AWS CLI offers an aws ecr get-login command that simplifies the login process.

fn get_download_url_for_layer(
&self,
input: GetDownloadUrlForLayerRequest
) -> RusotoFuture<GetDownloadUrlForLayerResponse, GetDownloadUrlForLayerError>

Retrieves the pre-signed Amazon S3 download URL corresponding to an image layer. You can only get URLs for image layers that are referenced in an image.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

fn get_lifecycle_policy(
&self,
input: GetLifecyclePolicyRequest
) -> RusotoFuture<GetLifecyclePolicyResponse, GetLifecyclePolicyError>

Retrieves the specified lifecycle policy.

fn get_lifecycle_policy_preview(
&self,
input: GetLifecyclePolicyPreviewRequest
) -> RusotoFuture<GetLifecyclePolicyPreviewResponse, GetLifecyclePolicyPreviewError>

Retrieves the results of the specified lifecycle policy preview request.

fn get_repository_policy(
&self,
input: GetRepositoryPolicyRequest
) -> RusotoFuture<GetRepositoryPolicyResponse, GetRepositoryPolicyError>

Retrieves the repository policy for a specified repository.

fn initiate_layer_upload(
&self,
input: InitiateLayerUploadRequest
) -> RusotoFuture<InitiateLayerUploadResponse, InitiateLayerUploadError>

Notify Amazon ECR that you intend to upload an image layer.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

fn list_images(
&self,
input: ListImagesRequest
) -> RusotoFuture<ListImagesResponse, ListImagesError>

Lists all the image IDs for a given repository.

You can filter images based on whether or not they are tagged by setting the tagStatus parameter to TAGGED or UNTAGGED. For example, you can filter your results to return only UNTAGGED images and then pipe that result to a BatchDeleteImage operation to delete them. Or, you can filter your results to return only TAGGED images to list all of the tags in your repository.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

List the tags for an Amazon ECR resource.

fn put_image(
&self,
input: PutImageRequest
) -> RusotoFuture<PutImageResponse, PutImageError>

Creates or updates the image manifest and tags associated with an image.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

fn put_lifecycle_policy(
&self,
input: PutLifecyclePolicyRequest
) -> RusotoFuture<PutLifecyclePolicyResponse, PutLifecyclePolicyError>

Creates or updates a lifecycle policy. For information about lifecycle policy syntax, see Lifecycle Policy Template.

fn set_repository_policy(
&self,
input: SetRepositoryPolicyRequest
) -> RusotoFuture<SetRepositoryPolicyResponse, SetRepositoryPolicyError>

Applies a repository policy on a specified repository to control access permissions.

fn start_lifecycle_policy_preview(
&self,
input: StartLifecyclePolicyPreviewRequest
) -> RusotoFuture<StartLifecyclePolicyPreviewResponse, StartLifecyclePolicyPreviewError>

Starts a preview of the specified lifecycle policy. This allows you to see the results before creating the lifecycle policy.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Adds specified tags to a resource with the specified ARN. Existing tags on a resource are not changed if they are not specified in the request parameters.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Deletes specified tags from a resource.

fn upload_layer_part(
&self,
input: UploadLayerPartRequest
) -> RusotoFuture<UploadLayerPartResponse, UploadLayerPartError>

Uploads an image layer part to Amazon ECR.

This operation is used by the Amazon ECR proxy, and it is not intended for general use by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Loading content...