Crate rusoto_ecr

Amazon EC2 Container Registry (Amazon ECR) is a managed AWS Docker registry service. Customers can use the familiar Docker CLI to push, pull, and manage images. Amazon ECR provides a secure, scalable, and reliable registry. Amazon ECR supports private Docker repositories with resource-based permissions using AWS IAM so that specific users or Amazon EC2 instances can access repositories and images. Developers can use the Docker CLI to author and manage images.

If you're using the service, you're probably looking for EcrClient and Ecr.

Structs

AuthorizationData
An object representing authorization data for an Amazon ECR registry.

BatchCheckLayerAvailabilityRequest

BatchCheckLayerAvailabilityResponse

BatchDeleteImageRequest
Deletes specified images within a specified repository. Images are specified with either the imageTag or imageDigest.

BatchDeleteImageResponse

BatchGetImageRequest

BatchGetImageResponse

CompleteLayerUploadRequest

CompleteLayerUploadResponse

CreateRepositoryRequest

CreateRepositoryResponse

DeleteRepositoryPolicyRequest

DeleteRepositoryPolicyResponse

DeleteRepositoryRequest

DeleteRepositoryResponse

DescribeImagesFilter
An object representing a filter on a DescribeImages operation.

DescribeImagesRequest

DescribeImagesResponse

DescribeRepositoriesRequest

DescribeRepositoriesResponse

EcrClient
A client for the Amazon ECR API.

GetAuthorizationTokenRequest

GetAuthorizationTokenResponse

GetDownloadUrlForLayerRequest

GetDownloadUrlForLayerResponse

GetRepositoryPolicyRequest

GetRepositoryPolicyResponse

Image
An object representing an Amazon ECR image.

ImageDetail
An object that describes an image returned by a DescribeImages operation.

ImageFailure
An object representing an Amazon ECR image failure.

ImageIdentifier
An object with identifying information for an Amazon ECR image.

InitiateLayerUploadRequest

InitiateLayerUploadResponse

Layer
An object representing an Amazon ECR image layer.

LayerFailure
An object representing an Amazon ECR image layer failure.

ListImagesFilter
An object representing a filter on a ListImages operation.

ListImagesRequest

ListImagesResponse

PutImageRequest

PutImageResponse

Repository
An object representing a repository.

SetRepositoryPolicyRequest

SetRepositoryPolicyResponse

UploadLayerPartRequest

UploadLayerPartResponse

Enums

BatchCheckLayerAvailabilityError
Errors returned by BatchCheckLayerAvailability

BatchDeleteImageError
Errors returned by BatchDeleteImage

BatchGetImageError
Errors returned by BatchGetImage

CompleteLayerUploadError
Errors returned by CompleteLayerUpload

CreateRepositoryError
Errors returned by CreateRepository

DeleteRepositoryError
Errors returned by DeleteRepository

DeleteRepositoryPolicyError
Errors returned by DeleteRepositoryPolicy

DescribeImagesError
Errors returned by DescribeImages

DescribeRepositoriesError
Errors returned by DescribeRepositories

GetAuthorizationTokenError
Errors returned by GetAuthorizationToken

GetDownloadUrlForLayerError
Errors returned by GetDownloadUrlForLayer

GetRepositoryPolicyError
Errors returned by GetRepositoryPolicy

InitiateLayerUploadError
Errors returned by InitiateLayerUpload

ListImagesError
Errors returned by ListImages

PutImageError
Errors returned by PutImage

SetRepositoryPolicyError
Errors returned by SetRepositoryPolicy

UploadLayerPartError
Errors returned by UploadLayerPart

Traits

Ecr
Trait representing the capabilities of the Amazon ECR API. Amazon ECR clients implement this trait.