Struct rusoto_ecr::EcrClient

pub struct EcrClient { /* fields omitted */ }

A client for the Amazon ECR API.

Methods

impl EcrClient

pub fn new(region: Region) -> EcrClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> EcrClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,